[API Pole emploi] Move function definition to its own module
This client was the only one to be exposed via the __init__.py
file, making its import different from the others.
