--- itou/utils/apis/__init__.py
@@ -1,12 +1,0 @@
-from django.conf import settings
-
-from itou.utils.apis.pole_emploi import PoleEmploiRoyaumePartenaireApiClient
-
-
-def pole_emploi_partenaire_api_client():
-    return PoleEmploiRoyaumePartenaireApiClient(
-        settings.API_ESD["BASE_URL"],
-        settings.API_ESD["AUTH_BASE_URL"],
-        settings.API_ESD["KEY"],
-        settings.API_ESD["SECRET"],
-    )
