+++ itou/utils/apis/pole_emploi.py
@@ -5,6 +5,7 @@ import re
 import time
 
 import httpx
+from django.conf import settings
 from django.core.cache import caches
 from unidecode import unidecode
 
@@ -312,3 +313,12 @@ class PoleEmploiRoyaumeAgentAPIClient(BasePoleEmploiApiClient):
     ]
     REALM = "/agent"
     CACHE_API_TOKEN_KEY = "pole_emploi_api_agent_client_token"
+
+
+def pole_emploi_partenaire_api_client():
+    return PoleEmploiRoyaumePartenaireApiClient(
+        settings.API_ESD["BASE_URL"],
+        settings.API_ESD["AUTH_BASE_URL"],
+        settings.API_ESD["KEY"],
+        settings.API_ESD["SECRET"],
+    )
